Stats & bugfixes and more bugs pop

+++ Assets/PROD/Scripts/Battle/Abilities/Abiliity Effects/AbilityEffect.cs
@@ -9,7 +9,7 @@ public abstract class AbilityEffect {
 [Serializable]
 class DamageEffect : AbilityEffect {
     
-    [SerializeField] private int attackRatio;
+    [SerializeField] private float attackRatio;
     
     public override void Apply(Unit caster, Unit target) {
         BattleLogic.Attack(caster, target, attackRatio);
